re-structured Services folder
